ci: pin to version v3.1.0
The reason we are using the "ci" prefix for the commit message is because We do not want to actually trigger a new release here, take a look at DX-2123 for more details.
Now that `npm-token` is no longer a required secret, we can check the release logs and see that OIDC Trusted Publishing is being used.
The change to the `npm-token` requirement was made in the pinned SHA version 3.1.0 of semantic-release-action/typescript.

Ticket: DX-2243